+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased changes
 
+## 1.14.0
+
 - Add `Moar.Assertions.assert_that/2` without the `:from` assertion.
 - Add `Moar.String.to_case/2`.